Refactor Stick processing to one place
The stick state processing relied on iterating through the `StickConfig`s in
`KeyConfig` to preprocess them, and loop through the resulting `Action` mappings
in `KeyState`.

This behavior has been consolidated into a single location: `StickActions`. May
revisit this further as it touches both configurations and the hot loop of
action processing.

Speaking of the hot loop... I was tinkering with how the bit masks are
processed, and vectorized the state matching so it's now 30% faster on average
and 80% faster in the worst case (we miss entirely). This was already a fast bit
of the program, but I couldn't help myself.